feat: :hammer: update game
