+++ api/games.ts
@@ -86,3 +86,23 @@ export const gameValidate = async (
     data,
   })
 }
+
+/**
+ * update game
+ * @param id
+ * @param data
+ * @returns
+ */
+export const updateGame = async (
+  id: number,
+  data: FormData
+): Promise<AxiosResponse<unknown>> => {
+  return await backend({
+    url: `/game-projects/${id}`,
+    method: 'PATCH',
+    headers: {
+      'Content-Type': 'multipart/form-data',
+    },
+    data,
+  })
+}
